+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Style and lint updates for code beauty.
 - Uses more specific exception subclasses.
 - Added optionalality to menus. For example, you can now specify that you don't want a policy template by selecting the ```0: <None> option.
+- Detects parent recipes that will require an empty <version> tag (uses PlistReader) and puts one in.
 
 ### FIXED
 - Solves unhandled exception when an invalid entry is made during auto mode.